Revert "bugfix: "강화 확률 오류 수정""
