+++ src/controllers/player.controller.js
@@ -1,20 +1,9 @@
 import { prisma } from '../utils/prisma/index.js';
-import { Prisma } from "@prisma/client";
+import { Prisma } from '@prisma/client';
 import { throwError } from '../utils/error/error.handle.js';
 import accountService from '../services/account.service.js';
-import playerService from '../services/player.service.js';
-import { PICKUP_TYPE, PICKUP_AMOUNT, PICKUP_PRICE, PLAYER_RANKS, UPGRADE_RESULTS, } from '../utils/enum.js';
-import { 
-  UPGRADE_COST, 
-  MAX_UPGRADE_MATERIALS,
-  UPGRADE_SUCCESS_RATES,
-  UPGRADE_MATERIAL_BONUSES,
-  DOWNGRADE_RATES,
-  DESTRUCTION_RATES,
-  NEXT_RANK,
-  PREV_RANK,
-} from '../utils/constants.js';
-import { MESSAGE_UPGRADE_RESULT } from '../utils/messages.js';
+import { PICKUP_TYPE, PICKUP_AMOUNT, PICKUP_PRICE } from '../utils/enum.js';
+import { calculateValue, calculatePrice, calculatePickupRate } from '../utils/valuation.js';
 
 /**
  * 팀 편성 로직
@@ -120,6 +109,15 @@ export async function pickupPlayer(req, res, next) {
     // 계정 인증+인가
     await accountService.checkAccount(accountId, authAccountId);
 
+    // 이거 캐시 구현할때 그대로 쓰셔도 될듯합니다
+    // // 잔액 확인
+    // const cashLog = await prisma.cashLog.findFirst({
+    //   where: { accountId },
+    //   orderBy: { createAt: 'desc' },
+    // });
+    // if (!cashLog || cashLog.totalCash < PICKUP_PRICE * numPulls)
+    //   throw throwError('캐시 잔액이 부족합니다.', 402);
+
     // 뽑기권 확인
     if (!(pickupType in PICKUP_TYPE)) throw throwError('올바르지 않은 뽑기권 종류입니다.', 400);
     if (!(pickupAmount in PICKUP_AMOUNT)) throw throwError('올바르지 않은 뽑기 횟수입니다.', 400);
@@ -223,153 +221,6 @@ export async function pickupPlayer(req, res, next) {
 }
 
 /**
- * 선수 강화 로직
- * @param {*} req
- * @param {*} res
- * @param {*} next
- * @returns
- */
-export async function upgradePlayer(req, res, next) {
-  const accountId = +req.params.accountId;
-  const authAccountId = +req.account;
-  const targetId = +req.body.targedId; // 강화할 선수 rosterId
-  const materials = req.body.materials; // 강화재료로 소모할 선수들
-
-  try {
-    // 계정 인증+인가
-    await accountService.checkAccount(accountId, authAccountId);
-
-    // 파라미터 무결성 확인
-    if (!targetId || isNaN(targetId)) throw throwError('올바르지 않은 targetId 형식입니다.', 400);
-    if (!materials) materials = []; // 빈 강화재료 input 형식 통일
-    if (!Array.isArray(materials)) throw throwError('올바르지 않은 materials 형식입니다.', 400);
-    for (const material of materials) {
-      if (!material || isNaN(+material)) throw throwError('올바르지 않은 materials 형식입니다.', 400);
-    }
-
-    // 최대 강화재료 수 초과 여부 확인
-    if (materials.length > MAX_UPGRADE_MATERIALS) throw throwError(`강화재료는 최대 ${MAX_UPGRADE_MATERIALS}개만 사용 가능합니다.`, 400);
-
-    // 캐시 잔액 확인
-    const cashLog = await prisma.cashLog.findFirst({
-      where: { accountId: accountId },
-      select: { totalCash: true },
-      orderBy: { createAt: 'desc' },
-    });
-    if (!cashLog || cashLog.totalCash < UPGRADE_COST)
-      throw throwError('캐시 잔액이 부족합니다.', 402);
-
-    // 강화대상 선수 보유 여부 확인
-    const targetRoster = await prisma.roster.findFirst({
-      where: { 
-        accountId : accountId, 
-        rosterId : targetId,
-      },
-      select: {
-        rosterId: true,
-        rank: true,
-      },
-    });
-
-    if (!targetRoster) throw throwError('강화할 선수를 보유하고 있지 않습니다.', 404);
-
-    // 최대강화 도달 여부 확인
-    let targetRank = targetRoster.rank;
-    if (targetRank === PLAYER_RANKS.LEGENDARY) throw throwError('이미 최대로 강화된 선수입니다.', 400);
-
-    // 기본 강화 성공률
-    let successRate = UPGRADE_SUCCESS_RATES.get(targetRank);
-
-    // 강화재료 적용
-    for (const materialId of materials) {
-      const materialRoster = await prisma.roster.findMany({
-        where: { 
-          accountId : accountId, 
-          rosterId : +materialId,
-        },
-        select: {
-          rosterId: true,
-          rank: true,
-        },
-      });
-
-      // 선수 보유 여부 확인
-      if (!materialRoster) throw throwError('강화재료로 사용할 선수를 보유하고 있지 않습니다.', 400);
-
-      // 강화 보너스 성공률 적용
-      successRate += UPGRADE_MATERIAL_BONUSES.get(materialRoster.rank);
-    }
-    successRate = Math.min(successRate, 1);
-
-    // 강화 진행
-    let result = null;
-    await prisma.$transaction(
-      async (tx) => {
-        let successRandom = Math.random();
-
-        // 강화 성공시:
-        if (successRate > successRandom) {
-          // 선수 강화
-          result = UPGRADE_RESULTS.SUCCESS;
-          await tx.roster.update({
-            data: {
-              rank: NEXT_RANK.get(targetRank)
-            },
-            where: {
-              rosterId : targetId,
-            }
-          });
-
-        // 강화 실패시:
-        } else {
-          // 강화 페널티 판정
-          let penaltyRandom = Math.random();
-          // 파괴
-          if (DESTRUCTION_RATES.get(targetRank) > penaltyRandom) {  
-            result = UPGRADE_RESULTS.DESTROYED;
-            await tx.roster.delete({
-              where: {
-                rosterId : targetId,
-              }
-            });
-          // 등급 하락
-          } else if (DOWNGRADE_RATES.get(targetRank) > penaltyRandom) {  
-            result = UPGRADE_RESULTS.DOWNGRADE;
-            await tx.roster.update({
-              data: {
-                rank: PREV_RANK.get(targetRank)
-              },
-              where: {
-                rosterId : targetId,
-              }
-            });
-          // 페널티 없음
-          } else {
-            result = UPGRADE_RESULTS.FAILURE;
-          }
-        }
-
-        // 강화 재료 소모
-        for (const materialId of materials) {
-          await prisma.roster.delete({
-            where: { 
-              rosterId : +materialId,
-            },
-          });
-        }
-      },
-      {
-        isolationLevel: Prisma.TransactionIsolationLevel.Serializable,
-      },
-    );
-
-    return res.status(201).json({ message: MESSAGE_UPGRADE_RESULT.get(result) });
-  } catch (error) {
-    next(error);
-  }
-}
-
-/**
  * 선수 판매 로직
  * @param {*} req
  * @param {*} res
@@ -413,7 +264,7 @@ export async function sellPlayer(req, res, next) {
     if (!player) throw throwError('선수가 존재하지 않습니다.', 404);
 
     // 선수 가격
-    const price = await playerService.calculatePrice(playerService.calculateValue(player));
+    const price = calculatePrice(calculateValue(player));
 
     const result = await prisma.$transaction(async (tx) => {
       // 캐쉬 변동 내역
@@ -558,11 +409,11 @@ export async function getAllPlayers(req, res, next) {
     const result = [];
     for (const player of playerList) {
       // 선수 가치
-      const value = await playerService.calculateValue(player);
+      const value = calculateValue(player);
       // 선수 가격
-      const price = await playerService.calculatePrice(value);
+      const price = calculatePrice(value);
       // 선수 뽑기
-      const pickupRate = await playerService.calculatePickupRate(value);
+      const pickupRate = calculatePickupRate(value);
 
       result.push({
         playerName: player.playerName,
@@ -589,7 +440,7 @@ export async function getAllPlayers(req, res, next) {
  * @param {*} res
  * @param {*} next
  */
-export async function getLineup(req, res, next) {
+export async function getLinenup(req, res, next) {
   const accountId = +req.params.accountId;
   const authAccountId = +req.account;
 
@@ -654,9 +505,7 @@ export async function rosterPl(req, res, next) {
     });
 
     // 선수 가격
-    const price = await playerService.calculatePrice(
-      await playerService.calculateValue(rosterPlayer.player),
-    );
+    const price = calculatePrice(calculateValue(rosterPlayer.player));
 
     // 데이터 가공
     const result = {
