+++ src/routes/player.router.js
@@ -2,12 +2,11 @@ import express from 'express';
 import {
   createLineup,
   pickupPlayer,
-  upgradePlayer,
   sellPlayer,
   getPlayers,
   buyToken,
   getAllPlayers,
-  getLineup,
+  getLinenup,
   rosterPl,
 } from '../controllers/player.controller.js';
 import authMiddleware from '../middlewares/auth.middleware.js';
@@ -23,11 +22,6 @@ router.patch('/lineup/:accountId', authMiddleware, createLineup);
  * 선수 뽑기 API
  */
 router.post('/pickup/:accountId', authMiddleware, pickupPlayer);
-
-/**
- * 선수 강화 API
- */
-router.patch('/upgrade/:accountId', authMiddleware, upgradePlayer);
 
 /**
  * 선수 판매 API
@@ -52,7 +46,7 @@ router.get('/players', getAllPlayers);
 /**
  * 라인업 조회 API
  */
-router.get('/lineup/:accountId', authMiddleware, getLineup);
+router.get('/lineup/:accountId', authMiddleware, getLinenup);
 
 /**
  * 보유선수 상세보기 API
